ml_model/requirements.txt
torch==2.4.1
torchvision==0.19.1
scikit-learn==1.5.2
pillow==10.4.0
numpy==2.1.1
matplotlib==3.9.2
